move construct_constraint to utils
